move criterion to dev-dependencies
