+++ jolt-core/Cargo.toml
@@ -28,7 +28,6 @@ ark-serialize = { version = "0.4.2", default-features = false, features = [
 ] }
 ark-std = { version = "0.4.0" }
 clap = { version = "4.3.10", features = ["derive"] }
-criterion = { version = "0.5.1", features = ["html_reports"] }
 digest = "0.8.1"
 enum_dispatch = "0.3.12"
 fixedbitset = "0.5.0"
@@ -68,11 +67,18 @@ common = { path = "../common" }
 tracer = { path = "../tracer" }
 bincode = "1.3.3"
 
+[dev-dependencies]
+criterion = { version = "0.5.1", features = ["html_reports"] }
+
 [build-dependencies]
 common = { path = "../common" }
 
 [[bench]]
 name = "iai"
+harness = false
+
+[[bench]]
+name = "sumcheck"
 harness = false
 
 [lib]
