+++ jolt-core/src/lib.rs
@@ -15,8 +15,5 @@ pub mod lasso;
 pub mod msm;
 pub mod poly;
 pub mod r1cs;
-mod subprotocols;
+pub mod subprotocols;
 pub mod utils;
-
-// Benchmarks
-pub use crate::subprotocols::sumcheck::bench::sumcheck_bench;
