+++ jolt-core/src/subprotocols/sumcheck.rs
@@ -1116,154 +1116,6 @@ impl<F: JoltField> SumcheckInstanceProof<F> {
     }
 }
 
-pub mod bench {
-    use super::*;
-    use crate::poly::dense_mlpoly::DensePolynomial;
-    use crate::poly::eq_poly::EqPolynomial;
-    use crate::subprotocols::sumcheck::{CubicSumcheckParams, SumcheckInstanceProof};
-    use crate::utils::index_to_field_bitvector;
-    use ark_bn254::Fr;
-    use ark_std::{rand::Rng, test_rng, UniformRand};
-    use criterion::black_box;
-
-    pub fn sumcheck_bench(
-        group: &mut criterion::BenchmarkGroup<'_, criterion::measurement::WallTime>,
-    ) {
-        let num_vars = 16;
-        let mut rng = test_rng();
-
-        // PLAIN
-        let r1 = vec![Fr::rand(&mut rng); num_vars];
-        let r2 = vec![Fr::rand(&mut rng); num_vars];
-        let r3 = vec![Fr::rand(&mut rng); num_vars];
-        let eq1 = DensePolynomial::new(EqPolynomial::new(r1).evals());
-        let left = DensePolynomial::new(EqPolynomial::new(r2).evals());
-        let right = DensePolynomial::new(EqPolynomial::new(r3).evals());
-        let params = CubicSumcheckParams::new_prod(
-            vec![eq1.clone()],
-            vec![left.clone()],
-            right.clone(),
-            num_vars,
-        );
-
-        let mut claim = Fr::zero();
-        for i in 0..num_vars {
-            let eval1 = eq1.evaluate(&index_to_field_bitvector(i, num_vars));
-            let eval2 = left.evaluate(&index_to_field_bitvector(i, num_vars));
-            let eval3 = right.evaluate(&index_to_field_bitvector(i, num_vars));
-
-            claim += eval1 * eval2 * eval3;
-        }
-
-        let coeffs = vec![Fr::one()];
-
-        group.bench_function("sumcheck unbatched 2^16", |b| {
-            b.iter(|| {
-                let mut transcript = ProofTranscript::new(b"test_transcript");
-                let params = black_box(params.clone());
-                let (_proof, _r, _evals) = SumcheckInstanceProof::prove_cubic_batched(
-                    &claim,
-                    params,
-                    &coeffs,
-                    &mut transcript,
-                );
-            })
-        });
-
-        // FLAGGED
-        let num_leaves = 1 << num_vars;
-        let mut vals1 = vec![Fr::rand(&mut rng); num_leaves];
-        let mut vals2 = vec![Fr::rand(&mut rng); num_leaves];
-        // Set approximately half of the leaves to 1
-        for _ in 0..num_vars / 2 {
-            let rand_index = rng.gen_range(0..num_leaves);
-            vals1[rand_index] = Fr::one();
-            vals2[rand_index] = Fr::one();
-        }
-        let poly_a = DensePolynomial::new(vals1);
-        let poly_b = DensePolynomial::new(vals2);
-        let r = vec![Fr::rand(&mut rng); num_vars];
-        let eq = DensePolynomial::new(EqPolynomial::new(r).evals());
-        let params = CubicSumcheckParams::new_prod(
-            vec![poly_a.clone()],
-            vec![poly_b.clone()],
-            eq.clone(),
-            num_vars,
-        );
-
-        let mut claim = Fr::zero();
-        for i in 0..num_vars {
-            let eval1 = poly_a.evaluate(&index_to_field_bitvector(i, num_vars));
-            let eval2 = poly_b.evaluate(&index_to_field_bitvector(i, num_vars));
-            let eval3 = eq.evaluate(&index_to_field_bitvector(i, num_vars));
-
-            claim += eval1 * eval2 * eval3;
-        }
-
-        let coeffs = vec![Fr::one()];
-
-        group.bench_function("sumcheck unbatched (ones) 2^16", |b| {
-            b.iter(|| {
-                let mut transcript = ProofTranscript::new(b"test_transcript");
-                let params = black_box(params.clone());
-                let (_proof, _r, _evals) = SumcheckInstanceProof::prove_cubic_batched(
-                    &claim,
-                    params,
-                    &coeffs,
-                    &mut transcript,
-                );
-            })
-        });
-
-        // BATCHED
-        let batch_size = 10;
-        let num_vars = 14;
-
-        let r_eq = vec![Fr::rand(&mut rng); num_vars];
-        let eq = DensePolynomial::new(EqPolynomial::new(r_eq).evals());
-
-        let mut poly_as = Vec::with_capacity(batch_size);
-        let mut poly_bs = Vec::with_capacity(batch_size);
-        for _ in 0..batch_size {
-            let ra = vec![Fr::rand(&mut rng); num_vars];
-            let rb = vec![Fr::rand(&mut rng); num_vars];
-            let a = DensePolynomial::new(EqPolynomial::new(ra).evals());
-            let b = DensePolynomial::new(EqPolynomial::new(rb).evals());
-            poly_as.push(a);
-            poly_bs.push(b);
-        }
-        let params =
-            CubicSumcheckParams::new_prod(poly_as.clone(), poly_bs.clone(), eq.clone(), num_vars);
-        let coeffs = vec![Fr::rand(&mut rng); batch_size];
-
-        let mut joint_claim = Fr::zero();
-        for batch_i in 0..batch_size {
-            let mut claim = Fr::zero();
-            for var_i in 0..num_vars {
-                let eval_a = poly_as[batch_i].evaluate(&index_to_field_bitvector(var_i, num_vars));
-                let eval_b = poly_bs[batch_i].evaluate(&index_to_field_bitvector(var_i, num_vars));
-                let eval_eq = eq.evaluate(&index_to_field_bitvector(var_i, num_vars));
-
-                claim += eval_a * eval_b * eval_eq;
-            }
-            joint_claim += coeffs[batch_i] * claim;
-        }
-
-        group.bench_function("sumcheck 10xbatched 2^14", |b| {
-            b.iter(|| {
-                let mut transcript = ProofTranscript::new(b"test_transcript");
-                let params = black_box(params.clone());
-                let (_proof, _r, _evals) = SumcheckInstanceProof::prove_cubic_batched(
-                    &joint_claim,
-                    params,
-                    &coeffs,
-                    &mut transcript,
-                );
-            })
-        });
-    }
-}
-
 #[cfg(test)]
 mod test {
     use super::*;
